+++ example-context/example-service/example-command/example-command-api/src/test/java/uk/gov/justice/services/example/cakeshop/command/api/MakeCakeCommandApiTest.java
@@ -2,6 +2,7 @@ package uk.gov.justice.services.example.cakeshop.command.api;
 
 
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.verify;
 import static uk.gov.justice.services.core.annotation.Component.COMMAND_API;
@@ -10,15 +11,18 @@ import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatc
 import static uk.gov.justice.services.test.utils.core.messaging.JsonEnvelopeBuilder.envelope;
 import static uk.gov.justice.services.test.utils.core.messaging.MetadataBuilderFactory.metadataWithDefaults;
 
-import uk.gov.justice.services.core.enveloper.Enveloper;
 import uk.gov.justice.services.core.sender.Sender;
+import uk.gov.justice.services.messaging.Envelope;
 import uk.gov.justice.services.messaging.JsonEnvelope;
-import uk.gov.justice.services.test.utils.core.enveloper.EnveloperFactory;
+import uk.gov.justice.services.messaging.spi.DefaultEnvelope;
+
+import javax.json.JsonObject;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
@@ -28,12 +32,11 @@ public class MakeCakeCommandApiTest {
     @Mock
     private Sender sender;
 
-    private Enveloper enveloper = EnveloperFactory.createEnveloper();
-
+    @InjectMocks
     private MakeCakeCommandApi commandApi;
 
     @Captor
-    private ArgumentCaptor<JsonEnvelope> envelopeCaptor;
+    private ArgumentCaptor<DefaultEnvelope> envelopeCaptor;
 
     @Test
     public void shouldHandleMakeCakeCommand() throws Exception {
@@ -45,14 +48,15 @@ public class MakeCakeCommandApiTest {
     @Test
     public void shouldHandleMakeCakeRequest() {
         commandApi = new MakeCakeCommandApi();
-        commandApi.enveloper = enveloper;
         commandApi.sender = sender;
 
         final JsonEnvelope envelope = envelope()
                 .with(metadataWithDefaults().withName("example.make-cake"))
                 .withPayloadOf("Field", "Value").build();
 
-        commandApi.handle(envelope);
+        final Envelope<JsonObject> jsonObjectEnvelope = commandApi.handle(envelope);
+
+        assertThat(jsonObjectEnvelope.payload().getString("status"), equalTo("Making Cake"));
 
         verify(sender).send(envelopeCaptor.capture());
         assertThat(envelopeCaptor.getValue().metadata().name(), is("example.command.make-cake"));